+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Fixed
+ - Custom Proton installations now use Steam Runtime installations when applicable
+
 ## [1.5.1] - 2021-05-10
 ### Fixed
  - bwrap containerization now tries to mount more root directories except those that have been blacklisted due to potential issues